requirements.txt
numpy
streamlit
pillow
keras
tensorflow-cpu
